requirements.txt
streamlit
pandas
numpy<2
matplotlib
scikit-learn
seaborn
openpyxl
joblib
Pillow

torch==2.2.1
torchvision==0.17.1
streamlit-lottie